+++ CHANGELOG.md
@@ -1,3 +1,52 @@
+2.4.0 (unreleased)
+
+- Support a Vimium new tab experience: the browser can be configured to open a blank Vimium page as
+  the new tab page. In Vimium's settings, the Vomnibar can configured to open on new tabs. See
+  [instructions](https://github.com/philc/vimium?tab=readme-ov-file#how-to-allow-vimium-to-work-on-new-tab-pages).
+  (https://github.com/philc/vimium/pull/4795)
+
+2.3.1 (2025-11-12)
+
+- Fix Vimium to work with Chrome 144. (https://github.com/philc/vimium/issues/4785)
+
+2.3.0 (2025-06-30)
+
+- Add a command listing page, which documents all commands and their options. Access it
+  [on the web](https://vimium.github.io/commands/), or from the Vimium Options page > Show available
+  commands.
+- Some internal CSS classes were changed for Vimium's UI. This may affect those who have customized
+  Vimium's CSS via the options page. (https://github.com/philc/vimium/issues/4668)
+- Breaking change: when creating a mapping for `setZoom`, a `level` argument is now required.
+  E.g.: `map z2 setZoom level=2.0`.
+- Make `Vomnibar.activateBookmark` accept a `query` option.
+  (https://github.com/philc/vimium/pull/4591)
+- Fix `openCopiedUrlInCurrentTab` doesn't launch search queries.
+  (https://github.com/philc/vimium/issues/4657)
+- Make `openCopiedUrlInCurrentTab` accept a `position` option.
+- Update `goPrevious` and `goNext` commands to handle google.com's new layout.
+  (https://github.com/philc/vimium/issues/4650)
+- Add a "hide update notifications" option for silencing "Vimium has been updated" notifications.
+  (https://github.com/philc/vimium/issues/4346)
+- Use dark mode styles in the HUD when the browser is in dark mode.
+- Bug fixes.
+
+2.2.1 (2025-03-20)
+
+- Fix findSelected and findSelectedBackwards commands (https://github.com/philc/vimium/issues/4655)
+- Fix openCopiedUrlInCurrentTab (https://github.com/philc/vimium/issues/4654)
+
+2.2.0 (2025-03-08)
+
+- Use the browser's default search engine. [(#2598)](https://github.com/philc/vimium/issues/2598)
+- Add "reload hard" command (R). ([#4445](https://github.com/philc/vimium/pull/4445)).
+- Add zoomIn (zi), zoomOut (zo), zoomReset (z0), and setZoom commands.
+  ([#4488](https://github.com/philc/vimium/pull/4488))
+- Add findSelected and findSelectedBackwards commands.
+  ([#4502](https://github.com/philc/vimium/pull/4502))
+- Options page: improve UI, add error validation.
+- Make tab commands handle Firefox hidden tabs.
+- Bug fixes.
+
 2.1.2 (2024-04-03)
 
 - Better fix for Vomnibar doesn't always list tabs by recency.
@@ -46,9 +95,10 @@
 
 2.0.1 (2023-10-04)
 
-- Fix exception when migrating some pre-v2.0 settings. ([#4323](https://github.com/philc/vimium/issues/4323))
+- Fix exception when migrating some pre-v2.0 settings.
+  ([#4323](https://github.com/philc/vimium/issues/4323))
 
-2.0.0 (2023-09-28 -- partially rolled out to users on the Chrome store)
+2.0.0 (2023-09-28)
 
 - Support manifest v3, as now required by Chrome. This involved a partial rewrite and many changes.
   Please report any new issues [here](https://github.com/philc/vimium/issues).
@@ -62,14 +112,15 @@
 - Allow find mode to work when using only private windows.
   ([#3614](https://github.com/philc/vimium/issues/3614))
 - Add a count option to closeTabsOnLeft and closeTabsOnRight commands, to allow binding a key to
-  "close just 1 tab on the left/right" rather than closing all tabs, as is the default. E.g. `map cl
+  "close just 1 tab on the left/right" rather than closing all tabs, as is the default. E.g.
+  `map cl
   closeTabsOnLeft count=1`. ([#4296](https://github.com/philc/vimium/pull/4296))
 - Add search completions for Brave Search. ([#3851](https://github.com/philc/vimium/pull/3851))
 - Make regular expressions in find mode work again; other find mode improvements.
   ([#4261](https://github.com/philc/vimium/issues/4261))
 - Bug fixes. ([#3944](https://github.com/philc/vimium/pull/3944),
-[#3752](https://github.com/philc/vimium/pull/3752),
-[#3675](https://github.com/philc/vimium/pull/3675))
+  [#3752](https://github.com/philc/vimium/pull/3752),
+  [#3675](https://github.com/philc/vimium/pull/3675))
 
 1.67.7 (2023-07-12)